Revised NPC interactions and collision detection
Dialogue now gradually displays as intended and can be progressed forward by pressing the interact key again to skip to the end of the currently displaying line. The player is also no longer able to occupy the same position as an NPC, but is able to move through them, which will need to be resolved.